elos/rpnfilter: Extend RpnFilter with stack size tracking for steps
Currently the RpnFilter does not check the size of its stack before
adding steps that modify the stack. This can lead to crashes when
there are not enough values on the stack for a given command,
e.g. "1 1 AND EQ", "1 AND" and so on.

This patch solves this issue by introducing a simple stack size
tracking value that is updated and checked for every token during
the filter string compilation stage, failing elosRpnFilterCreate()
with an error message describing the issue.

EventFilter and ProcessFilter are also affected by this patch
and will work without changes.