elos/eventbuffer: Increase default buffer size from 8 to 128
A default size of 8 is a good compromise for a safuVec
that will automatically grow once it reaches its limit,
but not really for a EventBuffer that starts dropping Events
once the limit is reached.

The default limit is increased to 128 for now,
though further tweaking of the value could be necessary.