elos/clientmanager: Remove unit test case 'ClientManagerActive' in elosClientManagerStop
This unit test case currently occasionally breaks
due to a race condition that causes the mocking (and the test itself) to fail.

Fixing this will need a deep dive into the ClientManager itself,
which is marked for an overhaul (as well as new features) in the near future.
As this will make a complete rewrite of the test necessary,
the code has been removed to prevent code rot.