elos/plugin: Implement support for function overloading
This patch extends elosPluginParam_t by a set of optional
function overrides. These can be either alternative names
for the functions (which are solved by dlsym as usual) or
direct pointers to custom functions.

Very useful for unit testing and likely be useful too when
we switch the ClientManager over later on.