elos/plugin: Implement alternative Initialize mode directly using a file
This patch introduces the capability of setting a file name directly,
as opposed to the current way of setting the search path and getting
the file name from a samconfConfig_t object. It is not exactly a
pretty solution, but will work for now until we get around to our
planned plugin design revisit.