elos/plugin: Add basic unit test for Load function
This is a very basic unit test proving that the function name overrides
work and not much else, resulting in poor coverage. Getting the coverage
up will be time consuming and is dependend on future samconf features;
At the same time the planned revisit of the plugin component will likely
also affect this test, so completing it will be postponed until
the mentioned issues are resolved.