elos/logaggregator: Remove success test case in unit test for elosLogAggregatorFindEvents
The success case broke due to changes in the plugin component.
I currently do not see a quick fix here as the plugin was crowbared into
functioning for this test, which interferes with the normal cleanup routines.

The solution here is likely to use function overload mocking to
simulate the plugin, but this will likely be a bit more work -
Due to this the test is removed to prevent bitrot and
to keep things running.