elos/clientmanager: Use safuFlags_t in core functions
Replaces all uses of the uint32_t 'status' field with
an atomic safuFlags_t 'flags' field. This also removes
the 'lock' field as it was only used for accessing 'status'.
The GetStatus function was renamed and moved into
the connection section for the time being since it
is the only place where the function is still used right now.