elos/clientmanager: Overhaul implementation of ClientConnection
More or less completely reworks the ClientConnection code,
unfortunately this couldn't be split into smaller commits
as a point was reached where everything would completely fall
apart by breathing in its general direction.

This also isn't finished yet due to time constraints,
as there are still a few not-so-nice parts present; It is a massive
improvement over the previous code base though, it doesn't have
the race condition problems that plagued the predecessor and has
much clearer and better separated code, so any further fixes
and improvements should be a lot easier to do.