elos/rpnfilter: Compile and run regular expressions during runtime
Currently REGEX commands with repitition markers (e.g. .* or .+)
will fail during runtime. It looks like this is caused by copying
the regex_t data structure around during the RpnFilter creation.
Fully fixing this problem will cause a lot of code changes and is
quite time consuming. A simpler solution for now is to simply
compile the regex during runtime, which costs performance but keeps
everything operational until we implement the real fix.

The change has no measurable impact on the round trip times when
testing with a handful of messages per second, as most of the delay
is caused by scheduling here - It will likely make a big difference
though when thousands of messages are handled per second.