scanner: Add shmem implementation
Initial implementation of an shmem scanner that
reads from a shared memory ringbuffer and converts
the log entries there into elosEvents.

There were some liberties taken with the implementation
as we had limited information to work with,
so some sections of the scanner are expected to change
in the future (most prominent being the locking mechanism
as well as the ringbuffer related parts of the code).