elos/plugin: Merge elosPluginWorker_t into elosPlugin_t
The former had barely anything left after the elosPluginContext_t
data split, so we move it back into elosPlugin_t. Uses safuFlags_t
instead of a boolean since we want to use safuFlags_t anyway in the
near future, so we might as well start now and save work later.