elos: Introduce libelosplugin
Move the shared library side of the headers in the plugin component
into the new libelosplugin library. This is not much right now,
but it is a start.

Also adds a couple of very simple helper functions to wrap some
of the more unintuitive pieces of plugin handling. More functions
will likely be added here in the future.