connectionmanager: Fix double call to elosConnectionManagerDeleteMembers.
As it is intended that a ConectionManager is stopped before it gets
deleted remove the call to ConectionManagerStop from
elosConnectionManagerDeleteMembers and avoid errors caused by stopping an
already stopped ConectionManager.
We do not add a check to elosConnectionManagerDeleteMembers if it is
running, as we should remove complexity and simply fail. We should not
try to catch each corner case.