Add support for pkg-config
Adds the cmake function and template file neccessary to automatically
generate and install a pkg-config file along the library files.

Also adds sections that use the generator appropriately.

This automatically uses the versions of the dependencies detected
with cmake's find_package() command, which keeps the "built with"
and "depends on" versions of the dependencies in sync.
json-c is the exception to the rule here with a fixed version,
as checking the version with find_package() causes cmake errors.