smoketest: Move common environment variables into separate file
It has become a quite common occurence that smoketest_interactive.sh
breaks due to environment variables not being properly updated when
the smoketest has been changed or was affected by other changes.

Moving the environment variables that are the same in smoketest.sh
and smoketest_interactive.sh into a separate file is a quick and easy
fix for this problem.