smoketest: Run executable test of compile_program_with_cpp only on Release build
This smoketest also tests the generated executable to make sure
function pointers are correctly handled. This however can cause
problems when ASAN is linked in when trying to run the executable.
Since ASAN is automatically disabled for Release builds we change
the smoketest to only run the binary test there.