ci: Add `--no-mocks` option to install_deps.
After recent updates to safu and samconf, the build and installation of
mock libraries must be explicit disabled. Former it was done implicitly
by disabling the build of the unit test. This in turn pulls in
dependencies to `cmocka_extensions` and `cmocka_mocks` which is not desired
in any case.
This adds a new option `--no-mocks` to `install_deps.sh` to disable
building all the `libmock_<projectname>`.