test/integartion: Add wait with timeout to elos connect keyword.
Until now it was not guaranteed that elosd is already ready and listening
on a TCP port , when tests starts to run. So this could lead to
unpredictable test behavior.
Now we wait for a given period of time and observe /proc/pid/net/tcp if
the requested port is opened or fail after a timeout.