test/integration: Remove obsolete 'elos is running' test.
This test was one of the first test cases. As we now frequently start
and stop elosd during the tests and can not determine the point in time
when this test is running. The test is only meaningful if it would be
executed directly after reboot, thus we can drop the test. We now have
enough other test which starts elos and verify that it is running as
expected.