test/integration: Implement new keyword to check for stored events.
With the new keyword the test shall be more robust against timing
issues. The test case for blacklist filtering fails sometimes as the
lookup via `elosc -f` occurs between the storage of the first and
second event. The new keyword allows to retry transparently the fetch
command until either a timeout or the expected amount of events are
seen.