cmake: Make sure to remove INSTALL_INTERFACE in install_component_headers
When we use install_component_headers and the components use the common
BUILD_INTERFACE and INSTALL_INTERFACE generator expression for the
include directories, the INSTALL_INTERFACE expressions will evaluate to
`/` what cause installing the current rootfs. This way we shall be save
to use INSTALL_INTERFACE generator expressions in our components.