test/plugin/tcp: Remove test cases for elosConfigGetElosdConnectionLimit.
The two test cases verifies that invalid values for the compile
constants ELOSD_CONNECTION_LIMIT  and invalid values in the environment
variable are handled correctly. The function
elosConfigGetElosdConnectionLimit uses elosPluginConfigGetInt which in
turn shall verify the correct behavior by itself. So the test is
superfluous at this point.

Another issue is that, at the moment it is not possible to mock
samconfConfigGetInt32 as before. This is due to the fact, that it is
used in libelosplugin and this is not the DSO under test so the
linker-wrapping will not work here.