smoketest: Add cmake file to install smoketests if desired.
Do not install smoketest_interactive.sh and docu files.
The smoketest_interactive.sh script is meant for use during development
and for convenience it uses bash and setup the environment to find
compiled binaries and libraries. It has no value if elos is properly
installed. In this case it is an error if the smoke test is unable to
find required dependencies.