plugincontrol: Do not store events on publishing.
Storing events is currently a dedicated call and must be done separately
by the plugins.
The problem here was a historical reason, that the TCP code understand
publish and store as one task and for the scanner this was two separate
tasks. So as we have the elosPluginStore  API we should use it, and it
is easier to change one plugin then 4 scanner plugins.