feature(plugins/syslog): implement systemd socket activation
Allow integration as a syslog daemon in systemd-based environments.

systemd-journald can forward messages to syslog daemon via using the
syslog protocol over a UNIX socket. When using this feature, the socket
is owned and opened by systemd and passed to the syslog daemon using the
socket activation mechanism.

Allow elosd to be used as a syslog daemon in systemd-based environments
by using the passed syslog socket.