fix(elosc): do not try TCP and Unix sockets if only one is requested
By default, elosc always attempts to connect to elosd using a unix
socket and then via TCP. This may lead to unexpected results (and smoke
test failures if an unrelated elosd is listening on the default socket
path).

Expected behaviour if only the connection options for TCP are specified
on the command line is that either TCP is tried first or that unix
sockets are not tried at all.

Fix this by not attempting to connect via a unix socket if only
TCP-related options are passed (and vice-versa), and only attempting
both options if either both or none are passed explicitly.