fix(CI): Ensure code_lint uses a more precise branch ref
The problem is that git cant distinguish between a branch and a file or
directory name if named equal. So `integration` as branch and directory
let git commands fail. No we use a more elaborate reference to specify
the branch name, hopefully we never have a directory
`refs/heads/<branch name>`.