fix(ci): extend docker-run to remove unused docker-integration-run
The docker-run.sh has the `-l|--link` parameter to be linked against
other containers. This is useful in case you want to start or run
manually the integration tests. Further more we can get rid of the
docker-integration-run.sh which was only referenced and used in the
docu for integration test.